Dummy Variable Coupling Removed (#3)
The project's major goal was to create models that predicted values using categorical variables rather than dummy variables. This resulted in declarative code, but there was a major issue: coupling. If you wished to use models, you had to make sure that the dummies inside the objects were properly labeled. This is a serious issue. Well, it was removed in this merge.

Updates:
- Only use numerical values. We add dummy variables support in Utils.py